Speichern in Benutzeroberfläche jetzt mit Abfrage ob, erlaubt ... (noch nicht bei Rename oder Löschen)

+++ Projectboard/ProjectBoardBasic/awinGeneralModules.vb
@@ -9822,6 +9822,8 @@ Public Module awinGeneralModules
         Dim anzahlNeue As Integer = 0
         Dim anzahlChanged As Integer = 0
         Dim DBtimeStamp As Date = Date.Now
+        Dim outputLine As String = ""
+
 
         Dim request As New Request(awinSettings.databaseURL, awinSettings.databaseName, dbUsername, dbPasswort)
 
@@ -9837,6 +9839,14 @@ Public Module awinGeneralModules
                     hproj.timeStamp = DBtimeStamp
                     If request.storeProjectToDB(hproj) Then
                         anzahlNeue = anzahlNeue + 1
+                    Else
+                        ' kann eigentlich gar nicht sein ... wäre nur dann der Fall, wenn ein Projekt komplett gelöscht wurde , aber der Schreibschutz nicht gelöscht wurde 
+                        If awinSettings.englishLanguage Then
+                            outputLine = "protected project: " & hproj.name & ", " & hproj.variantName
+                        Else
+                            outputLine = "geschütztes Projekt: " & hproj.name & ", " & hproj.variantName
+                        End If
+                        outPutCollection.Add(outputLine)
                     End If
                 Else
                     ' ein in dem Szenario enthaltenes Projekt wird gespeichert , wenn es Unterschiede gibt 
@@ -9848,7 +9858,11 @@ Public Module awinGeneralModules
                             ' alles ok
                             anzahlChanged = anzahlChanged + 1
                         Else
-                            Dim outputLine As String = "Fehler bei Speichern Projekt:" & vbLf & hproj.name & ", " & hproj.variantName
+                            If awinSettings.englishLanguage Then
+                                outputLine = "protected project: " & hproj.name & ", " & hproj.variantName
+                            Else
+                                outputLine = "geschütztes Projekt: " & hproj.name & ", " & hproj.variantName
+                            End If
                             outPutCollection.Add(outputLine)
                         End If
                     End If
@@ -9862,38 +9876,61 @@ Public Module awinGeneralModules
         Try
             If request.storeConstellationToDB(currentConstellation) Then
             Else
-                Call MsgBox("Fehler in Schreiben Constellation " & currentConstellation.constellationName)
+                If awinSettings.englishLanguage Then
+                    outputLine = "Error when writing scenario: " & currentConstellation.constellationName
+                Else
+                    outputLine = "Fehler beim Schreiben Szanario: " & currentConstellation.constellationName
+                End If
+                outPutCollection.Add(outputLine)
+
             End If
         Catch ex As Exception
-            Throw New ArgumentException("Fehler beim Speichern der Portfolios in die Datenbank." & vbLf & "Datenbank ist vermutlich nicht aktiviert?")
+            If awinSettings.englishLanguage Then
+                outputLine = "Error when writing scenario - Database active?"
+            Else
+                outputLine = "Fehler beim Schreiben Szanario - Datenbank läuft?"
+            End If
+            Throw New ArgumentException(outputLine)
         End Try
 
 
         Dim tsMessage As String = ""
         If anzahlNeue + anzahlChanged > 0 Then
-            If menuCult.Name = ReportLang(PTSprache.deutsch).Name Then
+            If awinSettings.englishLanguage Then
                 tsMessage = "Zeitstempel: " & DBtimeStamp.ToShortDateString & ", " & DBtimeStamp.ToShortTimeString
             Else
                 tsMessage = "Timestamp: " & DBtimeStamp.ToShortDateString & ", " & DBtimeStamp.ToShortTimeString
             End If
 
         End If
-        Dim txtMsg As String
-        If menuCult.Name = ReportLang(PTSprache.deutsch).Name Then
-            txtMsg = "Gespeichert ... " & vbLf & _
-                    "Szenario: " & currentConstellation.constellationName & vbLf & vbLf & _
-                    "Anzahl neue Projekte und Projekt-Varianten: " & anzahlNeue.ToString & vbLf & _
-                    "Anzahl geänderte Projekte / Projekt-Varianten: " & anzahlChanged.ToString & vbLf & _
-                    tsMessage
+        
+        If awinSettings.englishLanguage Then
+            outputLine = "Stored ... " & vbLf & _
+                "Scenario: " & currentConstellation.constellationName & vbLf & vbLf & _
+                "Number new projects/project-variants: " & anzahlNeue.ToString & vbLf & _
+                "Number changed projects/project-variants: " & anzahlChanged.ToString & vbLf & _
+                tsMessage
         Else
-            txtMsg = "Stored ... " & vbLf & _
-                    "Scenario: " & currentConstellation.constellationName & vbLf & vbLf & _
-                    "Number of new projects and project-variants: " & anzahlNeue.ToString & vbLf & _
-                    "Number of changed projects and project-variants: " & anzahlChanged.ToString & vbLf & _
-                    tsMessage
+            outputLine = "Gespeichert ... " & vbLf & _
+                "Szenario: " & currentConstellation.constellationName & vbLf & vbLf & _
+                "Anzahl neue Projekte und Projekt-Varianten: " & anzahlNeue.ToString & vbLf & _
+                "Anzahl geänderte Projekte / Projekt-Varianten: " & anzahlChanged.ToString & vbLf & _
+                tsMessage
+        End If
+        outPutCollection.Add(outputLine)
+
+        Dim msgH As String = ""
+        Dim msgE As String = ""
+        If awinSettings.englishLanguage Then
+            msgH = "Store Scenario "
+            msgE = "Messages"
+        Else
+            msgH = "Szenario speichern "
+            msgE = "Messages"
         End If
 
-        Call MsgBox(txtMsg)
+        Call showOutPut(outPutCollection, msgH, msgE)
+
 
 
     End Sub
@@ -10048,22 +10085,29 @@ Public Module awinGeneralModules
 
                     variantProject = AlleProjekte.getProject(keyV)
                     baseProject = AlleProjekte.getProject(keyB)
-
+                    '
+                    ' Sonderbehandlung alter Fehler bei Variantenbildung: Custom-Fields wurde nicht aus Base-Variant übernommen 
                     If Not IsNothing(variantProject) And Not IsNothing(baseProject) Then
 
+                        ' Sonderbehandlung wegen ehemaligem Fehler, wo bei Varianten-Bildung die Custom-fields aus der Base-Variant nicht übernommen wurden 
                         If variantProject.getCustomFieldsCount = 0 And baseProject.getCustomFieldsCount > 0 Then
                             variantProject.copyCustomFieldsFrom(baseProject)
                             Dim zeitStempel As Date = variantProject.timeStamp
 
-                            ' jetzt löschen, dann speichern 
-                            If request.deleteProjectTimestampFromDB(pname, variantName, zeitStempel) Then
-                                ' all ok 
-                                If request.storeProjectToDB(variantProject) Then
-                                    ' alles ok 
-                                Else
+                            ' jetzt löschen, dann speichern ; wenn das löschen schiefgeht aufgrund Schreibschutz, dann geht auch das Speichern schief ... 
+                            If writeProtections.isProtected(keyV, dbUsername) Then
+                                ' kann nichts machen ...
+                            Else
+                                If request.deleteProjectTimestampFromDB(pname, variantName, zeitStempel) Then
+                                    ' all ok 
+                                    If request.storeProjectToDB(variantProject) Then
+                                        ' alles ok; jetzt  
+                                    Else
 
+                                    End If
                                 End If
                             End If
+                            
 
                         End If
 
@@ -10082,6 +10126,8 @@ Public Module awinGeneralModules
                 End If
 
                 ' Ende Sonderbehandlung  
+                ' 
+                '
 
                 Dim timeStampsToDelete As Collection = identifyTimeStampsToDelete(pname, variantName, keepAnzVersions)
 
@@ -10095,17 +10141,32 @@ Public Module awinGeneralModules
                             ' all ok 
                             anzDeleted = anzDeleted + 1
                         Else
-                            outputLine = "-->Fehler beim Löschen von " & pname & ", " & variantName & ", " & singleTimeStamp.ToShortDateString
+                            If awinSettings.englishLanguage Then
+                                outputLine = "-->Error deleting (protected?): " & pname & ", " & variantName & ", " & singleTimeStamp.ToShortDateString
+                            Else
+                                outputLine = "-->Fehler beim Löschen (geschützt?): " & pname & ", " & variantName & ", " & singleTimeStamp.ToShortDateString
+                            End If
+
                             outputCollection.Add(outputLine)
                         End If
 
                     Next
 
-                    outputLine = pname & " (" & variantName & "): " & anzDeleted & " TimeStamps gelöscht"
+                    If awinSettings.englishLanguage Then
+                        outputLine = pname & " (" & variantName & "): " & anzDeleted & " timestamps deleted"
+                    Else
+                        outputLine = pname & " (" & variantName & "): " & anzDeleted & " TimeStamps gelöscht"
+                    End If
+
                     outputCollection.Add(outputLine)
 
                 Else
-                    outputLine = pname & " (" & variantName & "): 0 TimeStamps gelöscht"
+                    If awinSettings.englishLanguage Then
+                        outputLine = outputLine = pname & " (" & variantName & "): 0 timestamps deleted"
+                    Else
+                        outputLine = pname & " (" & variantName & "): 0 TimeStamps gelöscht"
+                    End If
+
                     outputCollection.Add(outputLine)
                 End If
 
@@ -10134,8 +10195,14 @@ Public Module awinGeneralModules
                                 ' all ok 
                                 anzDeleted = anzDeleted + 1
                             Else
-                                outputLine = "-->Fehler beim Löschen von " & pname & ", " & variantName & ", " & kvp.Key.ToShortDateString
+                                If awinSettings.englishLanguage Then
+                                    outputLine = "-->Error deleting (protected?): " & pname & ", " & variantName & ", " & kvp.Key.ToShortDateString
+                                Else
+                                    outputLine = "-->Fehler beim Löschen (geschützt?): " & pname & ", " & variantName & ", " & kvp.Key.ToShortDateString
+                                End If
+
                                 outputCollection.Add(outputLine)
+
                             End If
 
                         Next
@@ -10146,9 +10213,18 @@ Public Module awinGeneralModules
 
                 Else
                     If variantName = "" Then
-                        outputLine = "Löschen verweigert:  " & pname & " - Szenarien: "
+                        If awinSettings.englishLanguage Then
+                            outputLine = "delete denied: " & pname & " - Scenarios: "
+                        Else
+                            outputLine = "Löschen verweigert:  " & pname & " - Szenarien: "
+                        End If
+
                     Else
-                        outputLine = "Löschen verweigert:  " & pname & " (" & variantName & ") " & " - Szenarien: "
+                        If awinSettings.englishLanguage Then
+                            outputLine = "delete denied: " & pname & " (" & variantName & ") " & " - Scenarios: "
+                        Else
+                            outputLine = "Löschen verweigert:  " & pname & " (" & variantName & ") " & " - Szenarien: "
+                        End If
 
                     End If
                     outputLine = outputLine & projectConstellations.getSzenarioNamesWith(pname, variantName)
@@ -10180,44 +10256,32 @@ Public Module awinGeneralModules
             ElseIf hproj.variantName <> variantName Then
                 Dim key As String = calcProjektKey(pname, variantName)
                 AlleProjekte.Remove(key)
+
             Else
-                If variantName = "" Then
+                ' es wird in Showprojekte und in AlleProjekte gelöscht, ausserdem auch auf der Projekt-Tafel 
+                
+                Dim key As String = calcProjektKey(pname, variantName)
 
-                    Call MsgBox("die Basis Variante kann nicht gelöscht werden")
+                Try
 
-                ElseIf hproj.variantName = variantName Then
-                    ' es wird die Standard-Variante aktiviert 
-                    Dim stdProj As clsProjekt
-                    Dim stdkey As String = calcProjektKey(pname, "")
+                    ' jetzt muss die bisherige Variante aus Showprojekte rausgenommen werden ..
+                    ShowProjekte.Remove(hproj.name)
 
-                    Dim key As String = calcProjektKey(pname, variantName)
+                    ' die gewählte Variante wird rausgenommen
+                    AlleProjekte.Remove(key)
 
-                    Try
-                        stdProj = AlleProjekte.getProject(stdkey)
+                    Call clearProjektinPlantafel(pname)
 
-                        ' jetzt muss die bisherige Variante aus Showprojekte rausgenommen werden ..
-                        ShowProjekte.Remove(hproj.name)
+                Catch ex As Exception
+                    If awinSettings.englishLanguage Then
+                        outputLine = "delete denied: " & pname & " (" & variantName & ") " & " - Scenarios: ""Error when deleting: " & pname & " (" & variantName & ")"
+                    Else
+                        outputLine = "delete denied: " & pname & " (" & variantName & ") " & " - Scenarios: ""Fehler beim Löschen: " & pname & " (" & variantName & ")"
+                    End If
+                    outputCollection.Add(outputLine)
+                End Try
 
-                        ' die gewählte Variante wird rausgenommen
-                        AlleProjekte.Remove(key)
-
-                        ' die Standard Variante wird aufgenommen
-                        ShowProjekte.Add(stdProj)
-
-                        Call clearProjektinPlantafel(pname)
-
-                        ' neu zeichnen des Projekts 
-                        Dim tmpCollection As New Collection
-                        Call ZeichneProjektinPlanTafel(tmpCollection, stdProj.name, hproj.tfZeile, tmpCollection, tmpCollection)
-
-
-                    Catch ex As Exception
-
-                    End Try
-
-                Else
-                    Call MsgBox("Fehler beim Löschen der Variante")
-                End If
+                
             End If
 
 
@@ -19877,27 +19941,5 @@ Public Module awinGeneralModules
         isMissingDefinitionOK = checkResult
 
     End Function
-
-
-    ''' <summary>
-    ''' zeigt die in der OutputCollection gesammelten Rückmeldungen in einem Fenster mit Scrollbar 
-    ''' </summary>
-    ''' <param name="outPutCollection"></param>
-    ''' <param name="header"></param>
-    ''' <param name="explanation"></param>
-    ''' <remarks></remarks>
-    Public Sub showOutPut(ByVal outPutCollection As Collection, ByVal header As String, ByVal explanation As String)
-        If outPutCollection.Count > 0 Then
-
-            Dim outputFormular As New frmOutputWindow
-            With outputFormular
-                .Text = header
-                .lblOutput.Text = explanation
-                .textCollection = outPutCollection
-                .ShowDialog()
-            End With
-
-        End If
-    End Sub
 
 End Module
